rename back plugin and move its config
