+++ packages/commercetools/theme/plugins/commercetools-config.js
